requirements.txt
pandas==2.1.4
beautifulsoup4==4.12.2
selenium==4.16.0
python-dotenv==1.0.0
webdriver-manager==4.0.1
pywin32  